(DROP) meta-lenovo: hr630: u-boot: start u-boot from memory
During development on real hardware, I don't want to flash a new
u-boot or phosphor version/build every time when I do a small change.
This is not just time consuming, but it might also reduce the lifetim
of the SPI chip.

Instead, I will just chainload u-boot via u-boot. To get this running,
u-boot needs to know that it needs to start at "0x83000000" (0x80000000
is the start address of the memory).

The decision for "0x83000000" as address is more because of "I only want
to make sure that it doesn't get overwritten".

The chainloading of the u-boot image in u-boot is fairly simple
(ast_eth1 is the dedicated NIC. The name might be eth1 if you have
u-boot 2019 already flashed):
```
setenv ethact ast_eth1
dhcp
tftp 0x83000000 tftp-server-v4:u-boot.bin
md 0x83000000 # make sure content is there
go 0x83000000
```

Signed-off-by: Tom Siewert <tom.siewert@hetzner.com>